Refactor, separating logic from gui, TestPlotter formatting
